+++ lib/shortcuts.rb
@@ -3,15 +3,13 @@ Magickly.dragonfly.configure do |c|
     process :convert, "-filter Gaussian -resize #{size}"
   end
   
-  # Dragonfly currently chokes on this command - see pull request #73:
-  # https://github.com/markevans/dragonfly/pull/73
-  #
-  # c.job :tilt_shift do |coefficients|
-  #   if coefficients == 'true'
-  #     coefficients = "2,-2,0.5"
-  #   end
-  #   
-  #   action = "\\( +clone -sparse-color Barycentric '0,0 black 0,%[fx:h-1] white' -function polynomial #{coefficients} \\) -compose Blur -set option:compose:args 15 -composite"
-  #   process :convert, action
-  # end
+  c.job :tilt_shift do |coefficients|
+    if coefficients == 'true'
+      # use default polynomial coefficients
+      coefficients = "2,-2,0.5"
+    end
+    
+    action = "\\( +clone -sparse-color Barycentric '0,0 black 0,%[fx:h-1] white' -function polynomial #{coefficients} \\) -compose Blur -set option:compose:args 15 -composite"
+    process :convert, action
+  end
 end